% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users.R
\name{by_map}
\alias{by_map}
\alias{users_by_insight_area}
\alias{users_by_key_issue}
\title{Insights Area & Key Issues}
\usage{
users_by_insight_area(tm)

users_by_key_issue(tm)
}
\arguments{
\item{tm}{Transformation map as returned by \code{transformation_map}.}
}
\value{
A function to apply to mixpanel data.
}
\description{
Get Insights Areas/Key issues each user visited and the number of
events triggered on each Insights Area/Key Area.
}
\examples{
\dontrun{
data(sample_data)

tm <- transformation_map()
users_by_insight_area(tm)(sample_data)

get_key_issues <- users_by_key_issue(tm)
get_key_issues(sample_data)
}

}
