% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timestamp.R
\name{convert_timestamp}
\alias{convert_timestamp}
\title{Convert Timestamp}
\usage{
convert_timestamp(ts)
}
\arguments{
\item{ts}{Vector or bare column name of timestamp (\code{time}).}
}
\description{
Convert the Mixpanel timestamp to an object of clas \code{POSIXct}.
}
