% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events.R
\name{events_by}
\alias{events_by}
\alias{events_by_country}
\alias{events_by_ts}
\alias{events_by_referrer}
\title{Count events}
\usage{
events_by_country(data, events = ".", sort = FALSE)

events_by_ts(
  data,
  events = ".",
  sort = FALSE,
  units = c("mins", "hours", "days", "months")
)

events_by_referrer(data, events = ".", sort = FALSE, vars = c("domain", "url"))
}
\arguments{
\item{data}{Mixpanel data, an object of class \code{mixpanel_list_data}
as returned by \code{\link{parse_mixpanel_data}} where \code{tidy} was 
set to \code{FALSE}.}

\item{events}{Regular expression of events to keep where \code{.}
keeps all events, passed to \link[base]{grepl}.}

\item{sort}{Whether to sort the countries by number of events.}

\item{units}{Passed to \link[base]{round}.}

\item{vars}{The level of granularity desired, the full url or the domain only.}
}
\value{
A \link[tibble]{tibble} with the country code and the number 
of events.
}
\description{
Count the number of events by various groups.
}
\examples{
\dontrun{
data(sample_data)

# by country
cns <- events_by_country(sample_data)
plot(cns)

# browse page events by hour
ts <- events_by_ts(
  sample_data, 
  events = "Browse page",
  units = "hours"
)
plot(ts)
}

}
