% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{get_formatted_mixpanel_data}
\alias{get_formatted_mixpanel_data}
\title{Formatted Event}
\usage{
get_formatted_mixpanel_data(
  from_date = Sys.Date() - 7,
  to_date = Sys.Date(),
  event = "Browse page",
  api_secret = Sys.getenv("MIXPANEL_API_SECRET")
)
}
\arguments{
\item{from_date}{Object of class \code{date} from which, and to which, to query 
for the event. Both are inclusive.}

\item{to_date}{Object of class \code{date} from which, and to which, to query 
for the event. Both are inclusive.}

\item{event}{The event or events that you wish to get data for, 
as a \code{vector} or a \code{list}.}

\item{api_secret}{Your API secret defaults to an environment variable 
(\code{MIXPANEL_API_SECRET}).}
}
\description{
Returns formatted data from the "Segmentation" endpoint.
}
\examples{
\dontrun{get_formatted_mixpanel_data()}

}
