% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/language.R
\name{language_change}
\alias{language_change}
\title{Language Changes}
\usage{
language_change(data, events = "Change Language")
}
\arguments{
\item{data}{Mixpanel data, an object of class \code{mixpanel_list_data}
as returned by \code{\link{parse_mixpanel_data}} where \code{tidy} was 
set to \code{FALSE}.}

\item{events}{Regular expression of events to keep where \code{.}
keeps all events, passed to \link[base]{grepl}.}
}
\description{
Assess from which to which languages user switch.
}
\examples{
\dontrun{
data(sample_data)
language_change(sample_data, events = ".")
}

}
