% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{mixpanel_data}
\alias{mixpanel_data}
\alias{get_mixpanel_data}
\alias{as_mixpanel_data}
\alias{get_mixpanel_data_by_day}
\alias{get_mixpanel_data_by_week}
\title{Get Data}
\usage{
get_mixpanel_data(
  from_date = Sys.Date(),
  to_date = Sys.Date(),
  limit = NULL,
  event = NULL,
  where = NULL,
  jsonl = NULL,
  read = TRUE,
  api_secret = Sys.getenv("MIXPANEL_API_SECRET"),
  verbose = TRUE
)

as_mixpanel_data(jsonl)

get_mixpanel_data_by_day(
  from_date = Sys.Date() - 3,
  to_date = Sys.Date(),
  limit = NULL,
  event = NULL,
  where = NULL,
  jsonl = NULL,
  read = TRUE,
  api_secret = Sys.getenv("MIXPANEL_API_SECRET"),
  verbose = TRUE
)

get_mixpanel_data_by_week(
  from_date = Sys.Date() - 7,
  to_date = Sys.Date(),
  limit = NULL,
  event = NULL,
  where = NULL,
  jsonl = NULL,
  read = TRUE,
  api_secret = Sys.getenv("MIXPANEL_API_SECRET"),
  verbose = TRUE
)
}
\arguments{
\item{from_date, to_date}{Object of class \code{date} from which, and to which, to query 
for the event. Both are inclusive.}

\item{limit}{the max number of events to be returned.}

\item{event}{The event or events that you wish to get data for, 
as a \code{vector} or a \code{list}.}

\item{where}{An expression to filter events by. 
See the \href{https://developer.mixpanel.com/docs/data-export-api#section-segmentation-expressions}{expression section} 
on the main data export API page.}

\item{jsonl}{If \code{NULL} a temp JSON file is created to temporarily hold the results, 
if a path is passed then the intermediate results are stored on specified file.
Unlike the temp file the latter is not deleted. 
This option may be ueful if querying a lot of events.}

\item{read}{Whether to read the data back from the the \code{jsonl} file.
If \code{FALSE} the function invisibly return \code{jsonl}.}

\item{api_secret}{Your API secret defaults to an environment variable 
(\code{MIXPANEL_API_SECRET}).}

\item{verbose}{Whether to print helpful messages to the console.}
}
\description{
Get raw Mixpanel data.
}
\section{Functions}{

\itemize{
  \item{\code{get_mixpanel_data} - Fetch Mixpanel data.}
  \item{\code{get_mixpanel_data_by_day}, \code{get_mixpanel_data_by_day} - Splits the call by day or week.}
  \item{\code{as_mixpanel_data} - Converts an existing \code{jsonl} file to an object of class \code{mixpanel_data}.}
}
}

\examples{
\dontrun{
datasets <- get_mixpanel_data_by_day() 
}

}
\seealso{
\href{https://developer.mixpanel.com/docs/exporting-raw-data#section-export-api-reference}{Official documentation} of the API call implmented here.
}
