% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{mixpanel_data_to_json}
\alias{mixpanel_data_to_json}
\alias{mixpanel_data_to_json.mixpanel_data}
\alias{mixpanel_data_to_multiple_json}
\alias{mixpanel_data_from_json_dir}
\alias{mixpanel_data_from_json}
\title{Mixpanel to JSON}
\usage{
mixpanel_data_to_json(mixpanel_data, file)

\method{mixpanel_data_to_json}{mixpanel_data}(mixpanel_data, file)

mixpanel_data_to_multiple_json(mixpanel_data, dir, n)

mixpanel_data_from_json_dir(dir)

mixpanel_data_from_json(file)
}
\arguments{
\item{mixpanel_data}{An object of class \code{mixpanel_data} as 
returned by code{\link{get_mixpanel_data}} and \code{\link{as_mixpanel_data}}.}

\item{dir, file}{Directory or JSON file to save or read JSON files.}

\item{n}{Size of batches.}
}
\description{
Save results of \code{\link{get_mixpanel_data}} or \code{\link{as_mixpanel_data}} as JSON.
}
\examples{
\dontrun{
md <- get_mixpanel_data()  
mixpanel_data_to_json(md)
}

}
