% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfe.R
\name{mixpanel_engagement}
\alias{mixpanel_engagement}
\title{Engagement}
\usage{
mixpanel_engagement(data, events = ".")
}
\arguments{
\item{data}{Mixpanel data, an object of class \code{mixpanel_list_data}
as returned by \code{\link{parse_mixpanel_data}} where \code{tidy} was 
set to \code{FALSE} or the output of \code{\link{mixpanel_latest_activity}}.}

\item{events}{Regular expression of events to keep where \code{.}
keeps all events, passed to \link[base]{grepl}.}
}
\value{
A \link[tibble]{tibble} of user ids and frequency.
}
\description{
Computes total engagement of user.
}
\examples{
\dontrun{
data(sample_data)
mixpanel_engagement(sample_data)
}

}
