% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfe.R
\name{mixpanel_frequency}
\alias{mixpanel_frequency}
\title{Frequency}
\usage{
mixpanel_frequency(
  data,
  events = ".",
  units = c("mins", "hours", "days", "months")
)
}
\arguments{
\item{data}{Mixpanel data, an object of class \code{mixpanel_list_data}
as returned by \code{\link{parse_mixpanel_data}} where \code{tidy} was 
set to \code{FALSE} or the output of \code{\link{mixpanel_latest_activity}}.}

\item{events}{Regular expression of events to keep where \code{.}
keeps all events, passed to \link[base]{grepl}.}

\item{units}{Passed to \link[base]{round} and applied to argument 
\code{since} as well.}
}
\value{
A \link[tibble]{tibble} of user ids and frequency.
}
\description{
Computes frequency at which user engages with the platform.
}
\examples{
\dontrun{
data(sample_data)
mixpanel_frequency(sample_data)
}

}
