% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfe.R
\name{mixpanel_latest_activity}
\alias{mixpanel_latest_activity}
\title{Latest Activity}
\usage{
mixpanel_latest_activity(data)
}
\arguments{
\item{data}{Mixpanel data, an object of class \code{mixpanel_list_data}
as returned by \code{\link{parse_mixpanel_data}} where \code{tidy} was 
set to \code{FALSE}.}
}
\value{
A \link[tibble]{tibble} of user ids and time at 
which they were last seen on the paltform.
}
\description{
Computes when a user was last active.
}
\note{
This is based on any kind of event being triggered
rather than the inaccurate log in event.
}
\examples{
\dontrun{
data(sample_data)
mixpanel_latest_activity(sample_data)
}

}
