% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfe.R
\name{mixpanel_recency}
\alias{mixpanel_recency}
\title{Recency}
\usage{
mixpanel_recency(
  data,
  since = Sys.time(),
  units = c("mins", "hours", "days", "months")
)
}
\arguments{
\item{data}{Mixpanel data, an object of class \code{mixpanel_list_data}
as returned by \code{\link{parse_mixpanel_data}} where \code{tidy} was 
set to \code{FALSE} or the output of \code{\link{mixpanel_latest_activity}}.}

\item{since}{Since when to compute recency.}

\item{units}{Passed to \link[base]{round} and applied to argument 
\code{since} as well.}
}
\value{
A \link[tibble]{tibble} of user ids and recency as 
\code{difftime} object.
}
\description{
Computes how recently a user was last active.
}
\examples{
\dontrun{
data(sample_data)
mixpanel_latest_activity(sample_data) \%>\% 
  mixpanel_recency()
}

}
