% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfe.R
\name{mixpanel_rfe}
\alias{mixpanel_rfe}
\title{Recency, Frequency, Engagement}
\usage{
mixpanel_rfe(
  data,
  units = c("mins", "hours", "days", "months"),
  since = Sys.time()
)
}
\arguments{
\item{data}{Mixpanel data, an object of class \code{mixpanel_list_data}
as returned by \code{\link{parse_mixpanel_data}} where \code{tidy} was 
set to \code{FALSE}.}

\item{units}{Passed to \link[base]{round} and applied to argument 
\code{since} as well.}

\item{since}{Since when to compute recency.}
}
\description{
Computes the RFE score of users.
}
