% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{parse_mixpanel_data}
\alias{parse_mixpanel_data}
\alias{parse_mixpanel_data.mixpanel_data}
\alias{parse_mixpanel_data2}
\alias{parse_mixpanel_data2.mixpanel_data}
\title{Parse Mixpanel Data}
\usage{
parse_mixpanel_data(mixpanel_data, dir = NULL, n = NULL)

\method{parse_mixpanel_data}{mixpanel_data}(mixpanel_data, dir = NULL, n = NULL)

parse_mixpanel_data2(mixpanel_data)

\method{parse_mixpanel_data2}{mixpanel_data}(mixpanel_data)
}
\arguments{
\item{mixpanel_data}{An object of class \code{mixpanel_data} as 
returned by code{\link{get_mixpanel_data}} and \code{\link{as_mixpanel_data}}.}

\item{dir}{Directory to hold JSON files, if \code{NULL} temp is created and 
subsequently deleted.}

\item{n}{Batch size to process, if \code{NULL} the function attempts to set
a sensible batch size.}
}
\description{
Parse results of \code{\link{get_mixpanel_data}} or \code{\link{as_mixpanel_data}}.
}
\examples{
\dontrun{
md <- get_mixpanel_data()  
parse_mixpanel_data(md)
}

}
