% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sample_data}
\alias{sample_data}
\title{Sample}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 17522 rows and 49 columns.
}
\source{
this package
}
\usage{
sample_data
}
\description{
Sample of two days of data, over 26,000 events.
}
\keyword{datasets}
