% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm.R
\name{transformation_map}
\alias{transformation_map}
\title{Transformation Map}
\usage{
transformation_map(type = c("list", "igraph"))
}
\arguments{
\item{type}{The object to return, see value section below.}
}
\value{
Depending on \code{type}, either a list containing 
\code{nodes} and \code{edges} or an object of class \code{igraph}.
}
\description{
Create Transformation Map as currently live on 
\href{https://intelligence.weforum.org/}{SIG}.
}
