% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users.R
\name{users_by_client}
\alias{users_by_client}
\title{Users Client}
\usage{
users_by_client(data)
}
\arguments{
\item{data}{Mixpanel data, an object of class \code{mixpanel_list_data}
as returned by \code{\link{parse_mixpanel_data}} where \code{tidy} was 
set to \code{FALSE}.}
}
\value{
A \link[tibble]{tibble} of the user, the client,
and the number of times each user has visited SIG from the 
client,useful if there is a tie (\code{n}).
}
\description{
Compute users most frequent client used.
}
