% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sessions.R
\name{users_events}
\alias{users_events}
\title{Events}
\usage{
users_events(data, events = ".", sort = FALSE)
}
\arguments{
\item{data}{Mixpanel data, an object of class \code{mixpanel_list_data}
as returned by \code{\link{parse_mixpanel_data}} where \code{tidy} was 
set to \code{FALSE}.}

\item{events}{Regular expression of events to keep where \code{.}
keeps all events, passed to \link[base]{grepl}.}

\item{sort}{Whether to sort the countries by number of events.}
}
\description{
Count number of events by users.
}
\examples{
\dontrun{
data(sample_data)
users_events(sample_data, events = "Browse page", sort = TRUE)
}

}
